Landing page with a Web view and an "Enter VR" button that oens the VR glasses message
