+++ app/src/hvr/res/layout/fragment_enter_vr.xml
@@ -5,6 +5,17 @@
     android:layout_height="match_parent"
     android:background="@color/midnight">
 
+    <com.google.android.material.appbar.MaterialToolbar
+        android:id="@+id/toolbar"
+        style="@style/Widget.MaterialComponents.Toolbar.Primary"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:elevation="0dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:popupTheme="@style/Theme.WolvicPhone.PopupOverlay" />
+
     <ImageView
         android:id="@+id/hvr_glass_image"
         android:layout_width="0dp"
@@ -31,22 +42,5 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@id/hvr_glass_image" />
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/button_open_web"
-        style="@style/Widget.MaterialComponents.Button.TextButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:fontFamily="monospace"
-        android:padding="24dp"
-        android:text="@string/wolvic_com_underlined"
-        android:textAllCaps="false"
-        android:textColor="@color/azure"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
