+++ app/src/hvr/res/layout/fragment_legal_document.xml
