+++ app/src/main/res/values-es-rES/strings.xml
@@ -2157,6 +2157,7 @@ the Select` button. When clicked it closes all the previously selected tabs -->
     <!-- This string is displayed in the 2D phone screen when the user starts the app without an available device.
      This message is shown until the user correctly setups the device. -->
     <string name="hvr_connect_glass_message">Conecte sus HUAWEI VR Glass para usar esta aplicación</string>
+    <string name="hvr_enter_vr">Iniciar VR</string>
 
     <!-- These strings are displayed in the privacy dialog on first launch. -->
     <string name="privacy_policy">Política de privacidad</string>
