+++ app/src/main/res/values/strings.xml
@@ -2192,6 +2192,7 @@ the Select` button. When clicked it closes all the previously selected tabs -->
     <!-- This string is displayed in the 2D phone screen when the user starts the app without an available device.
          This message is shown until the user correctly setups the device. -->
     <string name="hvr_connect_glass_message">Connect your HUAWEI VR Glass to use this app</string>
+    <string name="hvr_enter_vr">Enter VR</string>
 
     <!-- These strings are displayed in the privacy dialog on first launch. -->
     <string name="privacy_policy">Privacy policy</string>
